Renames loglines*.py -> strings*.py
The existing naming would conflict with the expectations of the
test runner / representer / analyzer tooling.
